fix: salah urutan migration cu

+++ Back-end/database/migrations/2024_08_12_022957_create_orderitems_table.php
@@ -13,13 +13,13 @@ return new class extends Migration
     {
         Schema::create('orderitems', function (Blueprint $table) {
             $table->id();
-            $table->integer('order_id');
-            $table->integer('menu_id');
+            $table->unsignedBigInteger('order_id');
+            $table->unsignedBigInteger('menu_id');
             $table->foreign('order_id')->references('id')->on('orders')->cascadeOnDelete();
             $table->foreign('menu_id')->references('id')->on('menus')->cascadeOnDelete();
             $table->integer('quantity');
             $table->decimal('unit_price', 10, 2);
-            $table->timestamps();  
+            $table->timestamps();
         });
     }
